requirements.txt
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.2
fuzzywuzzy==0.17.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
passlib==1.7.1
Pillow==5.3.0
pymongo==3.7.1
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
